+++ cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@openrouter/spawn",
-  "version": "0.7.6",
+  "version": "0.7.7",
   "type": "module",
   "bin": {
     "spawn": "cli.js"
