+++ cli/src/aws/main.ts
@@ -15,9 +15,9 @@ import {
   runServer,
   uploadFile,
   interactiveSession,
-  saveLaunchCmd,
 } from "./aws";
 import { resolveAgent } from "./agents";
+import { saveLaunchCmd } from "../history.js";
 import { runOrchestration } from "../shared/orchestrate";
 import type { CloudOrchestrator } from "../shared/orchestrate";
