+++ cli/src/daytona/main.ts
@@ -10,9 +10,9 @@ import {
   runServer,
   uploadFile,
   interactiveSession,
-  saveLaunchCmd,
 } from "./daytona";
 import { resolveAgent } from "./agents";
+import { saveLaunchCmd } from "../history.js";
 import { runOrchestration } from "../shared/orchestrate";
 import type { CloudOrchestrator } from "../shared/orchestrate";
