+++ cli/src/digitalocean/main.ts
@@ -11,9 +11,9 @@ import {
   runServer,
   uploadFile,
   interactiveSession,
-  saveLaunchCmd,
 } from "./digitalocean";
 import { resolveAgent } from "./agents";
+import { saveLaunchCmd } from "../history.js";
 import { runOrchestration } from "../shared/orchestrate";
 import type { CloudOrchestrator } from "../shared/orchestrate";
 import { logStep } from "../shared/ui";
