+++ cli/src/hetzner/main.ts
@@ -11,9 +11,9 @@ import {
   runServer,
   uploadFile,
   interactiveSession,
-  saveLaunchCmd,
 } from "./hetzner";
 import { resolveAgent } from "./agents";
+import { saveLaunchCmd } from "../history.js";
 import { runOrchestration } from "../shared/orchestrate";
 import type { CloudOrchestrator } from "../shared/orchestrate";
