+++ cli/src/sprite/main.ts
@@ -10,12 +10,12 @@ import {
   verifySpriteConnectivity,
   setupShellEnvironment,
   saveVmConnection,
-  saveLaunchCmd,
   runSprite,
   uploadFileSprite,
   interactiveSession,
 } from "./sprite";
 import { resolveAgent } from "./agents";
+import { saveLaunchCmd } from "../history.js";
 import { runOrchestration } from "../shared/orchestrate";
 import type { CloudOrchestrator } from "../shared/orchestrate";
